move useMiddleEllipsis to label component
