+++ apps/wallet/src/ui/app/components/transactions-card/TxnActionLabel.tsx
@@ -2,6 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Text } from '_app/shared/text';
+import { useMiddleEllipsis } from '_hooks';
+
+const TRUNCATE_MAX_LENGTH = 8;
+const TRUNCATE_PREFIX_LENGTH = 4;
 
 type TxnTypeProps = {
     address: string;
@@ -16,9 +20,16 @@ export function TxnTypeLabel({
     isTransfer,
     isSender,
 }: TxnTypeProps) {
+    const receiverAddress = useMiddleEllipsis(
+        address,
+        TRUNCATE_MAX_LENGTH,
+        TRUNCATE_PREFIX_LENGTH
+    );
     const transferLabel = isSender ? 'To' : 'From';
     const label = isTransfer ? transferLabel : 'Action';
-    const content = isTransfer ? address : moveCallFnName?.replace(/_/g, ' ');
+    const content = isTransfer
+        ? receiverAddress
+        : moveCallFnName?.replace(/_/g, ' ');
 
     return content ? (
         <div className="flex gap-1 break-all capitalize mt-1">
